cleanup class after methods moved
